Move tools into one file each
